+++ casual/casual-inbound-api/src/main/java/se/laz/casual/jca/inflow/work/CasualServiceCallWork.java
@@ -26,6 +26,7 @@ import se.laz.casual.network.protocol.messages.service.CasualServiceCallReplyMes
 import se.laz.casual.network.protocol.messages.service.CasualServiceCallRequestMessage;
 
 import java.util.UUID;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import static se.laz.casual.network.ProtocolVersion.VERSION_1_3;
@@ -44,6 +45,7 @@ public final class CasualServiceCallWork implements Work
     private CasualNWMessage<CasualServiceCallReplyMessage> response;
     private ServiceHandler handler = null;
     private final SpanId spanId;
+    private boolean workFailedUnexpectedly = false;
 
     public CasualServiceCallWork(UUID correlationId, CasualServiceCallRequestMessage message, boolean isTpNoReply, ProtocolVersion protocolVersion, SpanId spanId)
     {
@@ -72,7 +74,7 @@ public final class CasualServiceCallWork implements Work
     @Override
     public void release()
     {
-        /**
+        /*
          * Currently no way to stop a service lookup or call.
          * Transaction context with which this Work is started
          * is applied with timeout that lies outside this code.
@@ -102,6 +104,12 @@ public final class CasualServiceCallWork implements Work
         catch( ServiceHandlerNotFoundException e)
         {
             log.warning( ()-> "ServiceHandler not available for: " + message.getServiceName() );
+        }
+        catch( Throwable t)
+        {
+            workFailedUnexpectedly = true;
+            // This shouldn't happen with a well-behaved service handler. If it does, the handler that threw might need fixing.
+            log.log( Level.SEVERE, "An exception was thrown while handing a TPNOREPLY service call", t );
         }
     }
 
@@ -134,6 +142,14 @@ public final class CasualServiceCallWork implements Work
                         .setTransactionState( TransactionState.ROLLBACK_ONLY );
             log.warning( ()-> "ServiceHandler not available for: " + message.getServiceName() );
         }
+        catch( Throwable t)
+        {
+            workFailedUnexpectedly = true;
+            replyBuilder.setError( ErrorState.TPESYSTEM )
+                        .setTransactionState( TransactionState.ROLLBACK_ONLY );
+            // This shouldn't happen with a well-behaved service handler. If it does, the handler that threw might need fixing.
+            log.log( Level.SEVERE, "An exception was thrown while handing a service call", t );
+        }
         finally
         {
             CasualServiceCallReplyMessage reply = replyBuilder
@@ -165,4 +181,8 @@ public final class CasualServiceCallWork implements Work
         this.handler = handler;
     }
 
+    public boolean failedUnexpectedly()
+    {
+        return workFailedUnexpectedly;
+    }
 }